Support reading a state file of how much time users have used today
The time limit is meant to be cumulative for all sessions, and not allow
users to bypass the time limit after a session ends by starting a new session.
To handle this, implement support for reading the time used today from a
state file.

This implements reading the state file, but to write it we will need to
implement support for being a PAM session module as well.